requirements.txt
asgiref==3.3.4
backports.zoneinfo==0.2.1
cloudinary==1.25.0
cryptography==3.4.8
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.4
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
oauthlib==3.1.1
Pillow==8.2.0
psycopg2-binary
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
sqlparse==0.4.1